requirements.txt
future
numpy
scipy
matplotlib
scikit-learn
keras
tensorflow-gpu==1.4.0
pydot-ng
h5py
Pillow
